Add DUT hot pixels and some cuts
The hot pixels algorithm is included for the
DUT (highly inspired from the tele.cc).
The isolation criteria for the cluster is modified
in order to considered as isolated a cluster with
no other clusters in a 8-columns, 8-rows cross.
The ph difference between pixels in the common mode
subtraction, and therefore, to be considered the pixel
as signal, is increase to 20 (avoiding noisy pixels)